+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for various repo notes for Repository model
 - Support for various fields of ModuleMd unit
-- Introduced YumRepository.related_x_repository properties that can retrieve
+- Introduced YumRepository.get_x_repository methods that can retrieve
   related binary, debug and source repository
 
 ## [2.8.0] - 2021-03-17